+++ integrations/csharp/Oicana/ITemplate.cs
@@ -4,7 +4,7 @@ using Oicana.Inputs;
 using CompilationOptions = Oicana.Config.CompilationOptions;
 using ExportFormat = Oicana.Config.ExportFormat;
 
-namespace Oicana.Template;
+namespace Oicana;
 
 /// <summary>
 /// An Oicana template